+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation_macos/FVPNativeVideoView.m
@@ -22,4 +22,9 @@
   return [[AVPlayerLayer alloc] init];
 }
 
+- (BOOL)isPictureInPictureActive {
+  // PiP is not supported on macOS in this implementation.
+  return NO;
+}
+
 @end